move tests from separate file to Live/IncrementTest.php
